IAI ProtoProto tables
These tables will be used to feed to ChatGPT via their frontend interface (using the Browsing model). The goal is to test a very experimental version of what we'd love IAI to eventually become. This will help us test questions and see what kind of answers we can get back.